reduce Expose mana cost by 1, increase mana regen speed, restore up to threshold on room transition

+++ src/Thing.zig
@@ -174,7 +174,7 @@ mana: ?struct {
     max: i32,
     regen: ?struct {
         timer: utl.TickCounter,
-        max_threshold: usize,
+        max_threshold: i32,
     } = null,
 } = null,
 faction: Faction = .object,
